Update for Kafka 0.10.0
The Dockerfile FROM had been updated but the auto expand code was no
longer compatible. Rather than fixing up the exiting code I copied,
almost verbatim, the `generateAssignment` function from the
ReassignPartitionsCommand. The only change I made was to use a sequence
of strings for the topic list instead of a topics-to-move style JSON
string.